backend/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
Pillow==10.2.0
python-magic==0.4.27
flask-cors==4.0.0
pandas==2.2.0
numpy==1.26.4
sentence-transformers==2.5.1
torch==2.2.0
